requirements.txt
torch
numpy
pyyaml
gymnasium
matplotlib
pandas
seaborn
gym
libsumo
traci
tqdm
mlx